+++ src/openapi/specs/ControllerActionSpecs.php
@@ -58,6 +58,19 @@ class ControllerActionSpecs extends BaseSpecs
     }
 
     /**
+     * Returns a full qualified action name.
+     * 
+     * This is mainly used for {{yii\base\Controller::actions()}} definition.
+     *
+     * @return string Assuming the actionName is `update-row` then it will return `actionUpdateRow`.
+     * @since 3.3.1
+     */
+    public function getActionName()
+    {
+        return 'action'.Inflector::id2camel($this->actioName);
+    }
+
+    /**
      * {@inheritDoc}
      */
     public function getReflection()
@@ -65,7 +78,7 @@ class ControllerActionSpecs extends BaseSpecs
         if ($this->getActionObject() instanceof InlineAction) {
             // read data from: actionMethodName()
             $reflector = new ReflectionClass(get_class($this->getControllerObject()));
-            return $reflector->getMethod('action'.Inflector::id2camel($this->actioName));
+            return $reflector->getMethod($this->getActionName());
         }
 
         return new ReflectionClass(get_class($this->getActionObject()));